requirements.txt
amqp==2.6.0
billiard==3.6.3.0
celery==4.4.6
celery-redbeat==1.0.0
future==0.18.2
importlib-metadata==1.7.0
kombu==4.6.11
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
six==1.15.0
tenacity==6.2.0
vine==1.3.0
zipp==3.1.0
